third_party/blink/web_tests/fast/css/selector-text-escape-expected.txt

Test for Bug 34529 - [CSSOM] issues with cssText and selectorText

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

List items must be green. Otherwise some selectors are invalid.
class="&#001;_class"
class="&#031;_class"
class="0_class"
class="9_class"
class="-0_class"
class="-9_class"
class="--_class"
class="&#129;_class"
class="日本語_class"
class="𐐀_class"
class="azAZ09_-_class"
class="(/:@[~_class"
id="a.external:visited"
myattr="my"value\"
PASS rules[0].selectorText is expectedTexts[0]
PASS rules[0].cssText is expectedTexts[0] + ' { color: green; }'
PASS rules[1].selectorText is expectedTexts[1]
PASS rules[1].cssText is expectedTexts[1] + ' { color: green; }'
PASS rules[2].selectorText is expectedTexts[2]
PASS rules[2].cssText is expectedTexts[2] + ' { color: green; }'
PASS rules[3].selectorText is expectedTexts[3]
PASS rules[3].cssText is expectedTexts[3] + ' { color: green; }'
PASS rules[4].selectorText is expectedTexts[4]
PASS rules[4].cssText is expectedTexts[4] + ' { color: green; }'
PASS rules[5].selectorText is expectedTexts[5]
PASS rules[5].cssText is expectedTexts[5] + ' { color: green; }'
PASS rules[6].selectorText is expectedTexts[6]
PASS rules[6].cssText is expectedTexts[6] + ' { color: green; }'
PASS rules[7].selectorText is expectedTexts[7]
PASS rules[7].cssText is expectedTexts[7] + ' { color: green; }'
PASS rules[8].selectorText is expectedTexts[8]
PASS rules[8].cssText is expectedTexts[8] + ' { color: green; }'
PASS rules[9].selectorText is expectedTexts[9]
PASS rules[9].cssText is expectedTexts[9] + ' { color: green; }'
PASS rules[10].selectorText is expectedTexts[10]
PASS rules[10].cssText is expectedTexts[10] + ' { color: green; }'
PASS rules[11].selectorText is expectedTexts[11]
PASS rules[11].cssText is expectedTexts[11] + ' { color: green; }'
PASS rules[12].selectorText is expectedTexts[12]
PASS rules[12].cssText is expectedTexts[12] + ' { color: green; }'
PASS rules[13].selectorText is expectedTexts[13]
PASS rules[13].cssText is expectedTexts[13] + ' { color: green; }'
PASS successfullyParsed is true

TEST COMPLETE
